capturar excepciones

+++ src/main/java/com/fullstack/duocflix/service/MovieService.java
@@ -2,6 +2,7 @@ package com.fullstack.duocflix.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import com.fullstack.duocflix.model.Movie;
 import com.fullstack.duocflix.repository.MovieRepository;
 
@@ -25,14 +26,26 @@ public class MovieService implements IMovieService {
     }
 
     @Override
-    public Movie createMovie(Movie movie) {
+    public Movie createMovie(Movie movie) throws Exception {
+
+        String valid = isvalidMovie(movie);
+        if (!valid.isEmpty())
+            throw new Exception(valid);
+
         return movieRepository.save(movie);
     }
 
     @Override
-    public Movie updateMovie(Long id, Movie movie) {
+    public Movie updateMovie(Long id, Movie movie) throws Exception {
+
         if (movieRepository.existsById(id)) {
+
+            String valid = isvalidMovie(movie);
+            if (!valid.isEmpty())
+                throw new Exception(valid);
+
             movie.setId(id);
+
             return movieRepository.save(movie);
         } else {
             return null;
@@ -40,7 +53,30 @@ public class MovieService implements IMovieService {
     }
 
     @Override
-    public void deleteMovie(Long id) {
-        movieRepository.deleteById(id);
+    public boolean deleteMovie(Long id) {
+
+        if (movieRepository.existsById(id)) {
+
+            movieRepository.deleteById(id);
+
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    private String isvalidMovie(Movie movie) {
+        if (movie.getTitle() == null || movie.getTitle().isEmpty())
+            return "Título no ingresado";
+        if (movie.getYear() == 0)
+            return "Año no ingresado";
+        if (movie.getDirector() == null || movie.getDirector().isEmpty())
+            return "Director no ingresado";
+        if (movie.getGenre() == null || movie.getGenre().isEmpty())
+            return "Género no ingresado";
+        if (movie.getSynopsis() == null || movie.getSynopsis().isEmpty())
+            return "Sinópsis no ingresada";
+
+        return "";
     }
 }
